利用 WindowSystem 来管理窗口
这样可以实现使用 esc 来关闭窗口
